Refactor: generator -> start
