+++ src/main/java/github/Louwind/Features/impl/start/GenericFeatureStart.java
@@ -1,10 +1,10 @@
-package github.Louwind.Features.impl.generator;
+package github.Louwind.Features.impl.start;
 
 import com.google.gson.JsonDeserializationContext;
 import com.google.gson.JsonObject;
 import com.google.gson.JsonSerializationContext;
-import github.Louwind.Features.generator.FeatureStart;
-import github.Louwind.Features.generator.FeatureStartType;
+import github.Louwind.Features.start.FeatureStart;
+import github.Louwind.Features.start.FeatureStartType;
 import github.Louwind.Features.impl.init.FeatureStarts;
 import github.Louwind.Features.pool.FeaturePool;
 import github.Louwind.Features.util.FeaturesJsonHelper;
